#47 [MOVE] 뷰모델 screen 파일 분리
