+++ app/src/main/java/com/kpaas/plog/presentation/auth/screen/LoginScreen.kt
@@ -38,6 +38,7 @@ import com.kpaas.plog.core_ui.theme.body3Regular
 import com.kpaas.plog.core_ui.theme.title1Bold
 import com.kpaas.plog.core_ui.theme.title2Semi
 import com.kpaas.plog.presentation.auth.navigation.AuthNavigator
+import com.kpaas.plog.presentation.auth.viewmodel.LoginViewModel
 import com.kpaas.plog.util.UiState
 import com.kpaas.plog.util.toast
 import kotlinx.coroutines.flow.first
