+++ app/src/main/java/com/kpaas/plog/presentation/main/navigation/MainNavGraph.kt
@@ -13,7 +13,7 @@ import com.kpaas.plog.presentation.report.navigation.ReportNavigator
 import com.kpaas.plog.presentation.report.screen.ReportRoute
 import com.kpaas.plog.presentation.reward.navigation.RewardNavigator
 import com.kpaas.plog.presentation.reward.screen.RewardRoute
-import com.kpaas.plog.presentation.search.screen.SearchViewModel
+import com.kpaas.plog.presentation.search.viewmodel.SearchViewModel
 
 fun NavGraphBuilder.mainNavGraph(
     mainNavigator: MainNavigator,
