+++ app/src/main/java/com/kpaas/plog/presentation/navigator/PlogNavHost.kt
@@ -35,7 +35,7 @@ import com.kpaas.plog.presentation.reward.navigation.RewardNavigator
 import com.kpaas.plog.presentation.reward.navigation.rewardNavGraph
 import com.kpaas.plog.presentation.search.navigation.SearchNavigator
 import com.kpaas.plog.presentation.search.navigation.searchNavGraph
-import com.kpaas.plog.presentation.search.screen.SearchViewModel
+import com.kpaas.plog.presentation.search.viewmodel.SearchViewModel
 import com.kpaas.plog.presentation.splash.SplashScreen
 
 @Composable
