+++ app/src/main/java/com/kpaas/plog/presentation/report/screen/ReportScreen.kt
@@ -62,6 +62,7 @@ import com.kpaas.plog.core_ui.theme.button4Semi
 import com.kpaas.plog.core_ui.theme.title2Semi
 import com.kpaas.plog.domain.entity.ReportListEntity
 import com.kpaas.plog.presentation.report.navigation.ReportNavigator
+import com.kpaas.plog.presentation.report.viewmodel.ReportViewModel
 import kotlinx.coroutines.launch
 
 @Composable
