+++ app/src/main/java/com/kpaas/plog/presentation/report/viewmodel/MyReportViewModel.kt
@@ -1,4 +1,4 @@
-package com.kpaas.plog.presentation.report.screen
+package com.kpaas.plog.presentation.report.viewmodel
 
 import androidx.lifecycle.ViewModel
 import com.kpaas.plog.domain.entity.MyReportListEntity
